trading-platform/services/data-ingestion/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Data sources
yfinance==0.2.33
ccxt==4.1.56
newsapi-python==0.2.7
tweepy==4.14.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0

# Time series
pandas==2.1.3
numpy==1.26.2
ta==0.11.0
pytz==2023.3

# Message queue
aio-pika==9.3.1
pika==1.3.2

# Redis
redis==5.0.1
aioredis==2.0.1

# Monitoring
prometheus-client==0.19.0

# Utils
httpx==0.25.2
tenacity==8.2.3
structlog==23.2.0
python-json-logger==2.0.7

# Text analysis
textblob==0.17.1
feedparser==6.0.10

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0